+++ doc/whats_new.rst
@@ -47,6 +47,14 @@ Enhancements
   to :func:`xarray.Dataset.xsimlab.run`.
 - More consistent dictionary format for output variables in the xarray
   extension (:issue:`85`).
+- %-formatting and str.format() code has been converted into formatted string
+  literals (f-strings) (:issue:`90`).
+- :func:`~xsimlab.foreign` has been updated to get the original description
+  of a foreign variable (:issue:`91`)
+- The ``autodoc`` parameter of the :func:`xsimlab.process` decorator now allows
+  to automatically add an attributes section to the docstring of the class to
+  which the decorator is applied, using the metadata of each variable declared
+  in the class (:issue:`67`).
 
 Bug fixes
 ~~~~~~~~~